Refactor async client
This change refactors async client so that it is owned by the cluster
manager. Now users can just send requests and fire and forget if they
don't need to cancel. This makes a lot of the consumers much simpler.
The fire and forget functionality will be used for router request
shadowing (similar to the HTTP tracer).