optimize getRawSlices() (#33)
From looking at perf traces, getRawSlices() comes up quite a lot. We end up
constructing and destroying a lot of std::vector. This replaces that code with
a thin layer above evbuffer which requires passing in a C array to be filled.
This is not nearly as nice but avoids any heap allocations. Eventually if we
switch to C++14 we could use std::dynarray which would be nicer.