http conn man: do not eat data and trailers if response has started (#81)
Previously we assumed that if an upstream response started, the upstream
server no longer cares about further request data. This is not the case in
bidirectional streaming scenarios such as gRPC streaming requests and responses.
We need to relax this requirement. I believe the only change required here beyond
dropping the restriction is to fix retry state handling in the router, however
this is a scary change so there might be follow on fixes.

Note that both the HTTP/1.1 and HTTP/2 codecs should not raise any further data
frames if a stream is reset which should make this handling safe.