fix response before request handling (#82)
This is a regression from 2e53d0. During situations where we perform a
continuation, filters do not expect to get further request data if they
either finish the response or reset. In this case, we must drop further
request data. This still allows bidirectional streaming to take place.

As part of this investigation I noticed a potential issue in nghttp2
where we don't always send a complete response if we reset after sending
the response (for example a redirect with body or a 404). I filed an issue
with nghttp2 and will follow up with this separately.