async client refactor (#92)
We have increasing use cases where we want Envoy filters to be able to make 
direct HTTP calls but use the full router feature set (retries, etc.). We also want
all of the same stats. Previously, the HTTP async client didn't support most of
the router features and also had some code duplication. This change makes the
HTTP async client use the router directly so that all standard features are 
available via request headers directly.

In follow ups we will add the ability for the async client to write to an access log,
as well as for a constant route to be passed in, which will be higher
performance than having to use request headers that need to be parsed.