outlier ejection: implement consecutive 5xx ejection
This commit sets up the ability to automatically eject hosts based on
N consecutive 5xx responses. Hosts are ejected for an ejection interval
multiplied by the number of times the host has been ejected (so if a host
keeps getting ejected it will be ejected for longer and longer periods).

The ejection interval, number of consecutive 5xxs required for ejection,
and the maximum % of a cluster that can be auto ejected are all runtime
configurable.

This commit does not include documentation. The feature is going to evolve
a bit and we will do documentation at a later time when the feature is more
baked.

The main thing missing from this is the ability to log ejections/additions
so that external monitoring/alarming systems can visualize what is happening
since it's possible for auto ejection to be quite confusing if there isn't
enough information.