remove use of bufferevent for both plain and ssl connections
From perf investigations, a big slow down at high throughput is inside
libevent. bufferevent does not use edge triggered events, so it ends up
calling epoll_ctl on very write.

This change implements our own version of bufferevent inside of
ConnectionImpl. We need to do this for both plain and ssl connections.

This is an incredibly scary change. I've already run this on a prod box
for a while but will do more smoke testing before we merge and do a full
canary.