tcp proxy: connect to upstream when downstream connections (#175)
The previous behavior was to connect only when data is received on the
downstream connection. Unfortunately this breaks protocols which connect
and the server is the first half to send data (e.g., mysql).

This required a refactor to handle the case where we immediately disconnect
downstream during network filter chain creation.

This commit also cleans up the filter factory functions so that they take a
filter manager interface vs. a connection interface to force filters to
operate via the filter callbacks.