perf: new header map implementation (#209)
This change swaps out the current header map implementation for one that
is heavily optimized for performance. We implement our own string implementation
used for headers, as well as a new map implementation that allows direct access
to all headers that Envoy uses during fast path processing. We attempt to copy
and allocate as little as possible. There are some more wins that can be had that
are noted in the code but this change is too large as it is and we can do some
of them in follow ups.