tcp proxy: fix connection/fd leak during local close
This can happen when the connection gets called locally by another filter
(e.g., the rate limit filter).

This is a regression from both c172a4 and a10174. The first commit makes it
so that we create the upstream connection in the constructor and not when
we first have data. The second commit which removed bufferevent removes the
safeguard that would have happened in the destructor which would have closed
the fd in the bev.

This commit fixes the main leaking issue but also puts a failsafe into the
connection destructor such that we will close the fd always and be more likely
to crash instead of keep going.