network filters: add onNewConnection() callback
This is more fallout from c172a4. For network filters we want the
ability to do things when the connection is created, but also to
allow stopping iterations of further filters. In the previous change,
we made it so that the proxy filter will always make a new connection
to upstream regardless of what other filters do. We change that now
such that the tcp proxy filter will only make the upstream connection
in the onNewConnection() callback. Other filters (currently client_ssl
and ratelimit) now do their work on their own onNewConnection() callback
and stop iteration if necessary to prevent the tcp proxy filter from
making the upstream connection until it is supposed to.