network: cleanup connection logic and crash when fd cannot be allocated
The fix in 8d5366 looks good. This commit further cleans up the socket
close logic to make it simpler and more unified. It also changes the
behavior to crash the process if we can't allocate a socket or we have
an error during accept. In practice those only happen when we run out of
FDs which should be considered an OOM condition where we also crash on
purpose.