outlier detection: charge all failures to host
Previously we did only charged the final response code to a host and
ignored retires. For outlier detection purposes, we need to charge all
failures to the host to account for retries.

This commit also includes a small refactor to move upstream host into
the upstream request where it should be.