refactor to prepare for dynamic cluster add/remove (#299)
This change refactor is required so that we can support dynamic cluster
add and remove. Previously, we had an unfortunate circular dependency where
a host needs access to a bunch of information about the cluster. Switching
that over to shared pointers creates cyclic strong dependencies. Instead, we
refactor Cluster into Cluster and ClusterInfo, and now ClusterInfo is shared
between the primary cluster and hosts. This removes the cyclic dependency.

In follow up commits we will actually add the dynamic add/remove functionality
which requires some more plumbing changes, but this change is large enough as it
is.