redis: initial codec and proxy support
This is an initial implementation of redis support in Envoy. It
includes a codec as well a simple proxy filter and "connection pool."
The connection pool is trivial and just makes a single connection. It
does no consistent hashing, op splitting, etc. However, this commit
is sufficient for Envoy to proxy to a single redis server and
encode/decode the entire conversation. All of the main interfaces
are in place.

To keep this change as self contained as possible there is quite a bit
missing that will be done in individual follow up changes:
  - stats
  - timeouts
  - consistent hashing
  - supported op checking and splitting where appropriate (e.g., MGET)
  - A real connection pool that supports hashing, and M connections to
    N upstreams
  - redis active HC and outlier detection