local info: refactor information about the local env
This has been bothering me for a while. We are copying things like local
zone, local cluster, etc. all over the place. It's also likely that we will
add more things in the future. This introduces a LocalInfo abstract class
which holds such information.

The change is large by number of lines but it's simple.