stats: add scope concept
Currently stats are allocated from a fix memory pool in shared memory
so that we can do hot restart. We have no facility to delete stats. This
is not going to work when we can add/remove clusters dynamically. This
commit adds a new scope concept and plumbs through a scope for all cluster
stats. The actual scope implementation doesn't do anything currently other
than just pass through to the main store. The actual shared memory
implementation of deletion is quite complicated so I will do that in a
dedicated smaller change.

A few other notes:
1) I got rid of alt_stat_prefix from cluster. This was used a long time ago
   when we were rolling out sds. It's hard to support in the new scheme so
   I'm just getting rid of it.
2) How we manage SSL contexts is also broken for cluster deletion. The contexts
   are currently global and not deleted. I will fix this in a follow up.