stats: new thread local store implementation
This is needed to support dynamic add/remove of clusters and
transitively stat scopes.

There is no way around the fact that this is an extremely complicated
change. See the comment in thread_local_store.h for more info.

NOTE: This changes the shared memory layout and will require a full
      restart.