Fix handling of subnet "0.0.0.0/0" in whitelist.
"0.0.0.0/0" is a well-known subnet that contains all possible IPv4 addresses.
Envoy whitelist code currently does not handle the "/0" case correctly.

The code uses this formula to compute the subnet mask:
white_list_entry.ipv4_mask_ = ~((1 << (32 - mask)) - 1);

The LHS is an int32_t and the RHS is also computed as a 32-bit value.
However, the (<< 32) operation is invalid for a 32-bit value (undefined behavior).

The fix makes the RHS first compute as a 64-bits value that later gets
truncated to 32-bits on assignment.