Remove `readdir_r` in favor of `readdir`
`readdir_r` has been considered harmful for a while now and it's
deprecated by glibc since 2.24. I know `readdir_r` was historically the
thread-safe alternative but even if POSIX don't require `readdir` to be
thread-safe, it is thread-safe in any modern implementation such as
glibc.

Note that the docs says that the next POSIX spec will require `readdir`
to be thread-safe "officially".