orderly cm/cds/rds initialization
We have a fairly complicated init situation at this point in which we need to
do the following:
1) Initialize static cm clusters
2) Initialize cds cm clusters
3) Initialize rds
4) Start listening

This change sets up a generic init manager framework that allows extensions to
initialize after cm init but before we start listening.

This commit also adds "integration tests" for CDS/RDS in that we specify them
in one of the integration test configs. Nothing is currently implemented but it
at least checks that the server initialized correctly.

fixes https://github.com/lyft/envoy/issues/465