Handle `EAGAIN` when querying `SO_ERROR` on bsd
When a connect() is performed on a non-blocking socket, it'll return
-1/EINPROGRESS immediatly and when select()'ed for writability it'll
succeed or fail but when it fails the reason is hidden.
`getsockopt(,,SO_ERROR,,)` solves this issue on some sytems but on some
cases such as FreeBSD getsockopt returns 0 even on `EAGAIN`. The
following trick issues a `getpeername()` which will return -1/ENOTCONN
if the socket is not connected and read(,,1) will produce the right errno.