validation: Add config validation mode. (#499)
This adds a --mode command-line flag. "--mode=serve" is the default and
can be omitted; in that mode Envoy runs as it always has.

With "--mode=validate" Envoy checks that the configuration file in
--config-file is valid, then exits. "Valid" means that it performs the
JSON schema check as usual, but also initializes as much of its internal
state as possible, exiting with any errors as it would under
"--mode=serve". When initialization is complete, instead of listening
for traffic, Envoy prints an "OK" message to stderr and exits, returning
success.

In validation mode, to the extent possible, Envoy avoids interfering
with its environment. For example, the hot-restart process is skipped,
so validation can safely be run on the same machine as a serving Envoy.
Validation mode also doesn't send any upstream traffic or open any
listeners.

It *does* attempt to read any files referenced in the config, like
certs and private keys, so validation will fail if those files aren't
located at the expected paths. A future "lite validation" mode, that
mocks out the filesystem as well as the network, is not yet implemented
as of this patch.