fix SSL write crash when write buffer has a very large chain
Previously, we would get all buffer slices on the stack and stack overflow
if the chain was very large. This change limits the number of slices that we
write during each iteration. There are a number of potential improvements
possible in terms of collapsing small moves, enforcing fairness in terms of
number of iterations, etc.

This commit also removes a bunch of RawSlice to evbuffer_iovec conversions
for performance reasons.

NOTE: The added test repros the crash before the fix.

fixes https://github.com/lyft/envoy/issues/585