Fix config generation regression (#587). (#591)
Also added a "set -e" to configgen.sh to ensure we catch this in presubmit in the future. As a
bonus, plumbed in arbitrary address binding to admin server config, since we had some configs that
had been updated to this already and it seems a generally useful thing to have.