bazel: eliminate compiler relative paths from copts (#415). (#604)
Previously, when Envoy was included as an external dependency in the Istio build, the use of
compiler relative paths broke the build, since something like "-iquoteinclude" actually needed to be
"-iquoteexternal/envoy_git/include", and Bazel isn't able to transform raw copts.

This patch makes use of include_prefix to preserve the current Envoy #include idioms for include/
and source/ without having to resort to -iquote. The only drawback is we lose the ability to have
foo.cc refer to foo.h in something like source/common/common via #include "foo.h" - it must be
expressed as #include "common/common/foo.h". We will need to make these changes across the code base
as we Bazelify.

As a bonus, added some scripts to make it easier to deal with syncing BUILD files post #602.