GuardDog: Deadlock detection for stats/termination
This commit adds a deadlock detector called GuardDog. It replaces the
existing watchdog timer and performs the same statistic collecting
functionality while also providing the ability to detect and act on true
deadlocks (one or more event processing threads becoming permanently
nonresponsive). By performing all deadlock checking in an independent
thread we can PANIC() and kill the process for a semi-graceful restart
rather than hanging.  This should produce a core file which can be used
to find out details on the stuck threads.

New configuration variables are introduced to the top level
configuration to control the new kill behavior and also make the
existing thresholds for statistics incrementing configurable. The
configuration details are described in the configuration overview
document.

The new bits of implementation were placed into Event namespace because
they are most relevant to threads that are following the Event and
Dispatcher paradigm.  If this is not ideal it can be moved.