Backward: Add friendly stack trace capability.
Use the Backward-cpp project (https://github.com/bombela/backward-cpp)
to log stack traces on demand as desired for debugging and in response
to serious errors like PANIC() and assertion failures.

Unfortunately the best output from Backward-cpp requires significant
dependencies (elfutils, or worse GPL binutils) that we don't want to
pull in.  Instead, a Python script provides automated de-interleaving of
stack trace output from multiple threads and calls addr2line to decode
the addresses in the stack trace using the DWARF info present in the
binary.  The tool accepts Envoy log output on stdin or takes arguments
for a command to run as a child process and backtrace enhance the
output. For example, this command will run a single test under the
backtrace enhancer and show the output:

bazel test -c dbg //test/common/common:backwards_test
--run_under=`pwd`/tools/stack_decode.py
--strategy=TestRunner=standalone --cache_test_results=no
--test_output=all

Since this is intended only as a debugging tool the test care merely
exercises the code to ensure it does not crash without an affirmative
passing case.