ci: support building recipes in distinct --prefix directories (#415).
The plan is to build the dependencies for the developer-local flow from the same build recipes as
the CI image and using the same recursive make. This patch makes this is a bit easier, as one of the
things we need to do is figure out which headers each dependency exports for Bazel, and building
each recipe in its own --prefix makes this trivial.

This is a nop for the CI Docker image. Validated by building a CI Docker image and performing a
debug build/test against it.