ci: build and install custom Bazel binary for coverage runs.
This version of Bazel will only be used for code coverage purposes. It's necessary due to
the regression in C++ support noted in https://github.com/bazelbuild/bazel/issues/1118.

The plan is to use regular Bazel for build types other than coverage, since we want to make sure
that we don't depend on anything in this snowflake.

Also added lcov to the build, since Bazel coverage needs this to be available for generating
trace files from the .gcno/.gcda for each test.