ci: make bazel.debug build and test from a consuming project.
This allows us to catch in CI breakages such as those fixed in #732. Also, it allows us to validate
our example and story for other projects that consume Envoy for custom linking of filters keeps
working.

This adds the requirement of bind mounting in docker run with -v <some path>:/build for bazel.debug
and bazel.coverage builds.