ci: fix gcovr dependency path. (#769)
After #747, we no longer have the BUILD_DISTINCT=1 set of dependencies in the CI image, so fix the
gcovr path to solve the failures in #760.

Also fixed a snafu in configs/configgen.sh that I hit when running Docker as non-root. It was trying
to peek at the user's home directory, with no passwd entry.