test: use explicit openssl.cnf for generated certs.
Previously, openssl was picking up on whatever the default openssl.cnf path it was compiled with
pointed at, e.g. /etc/ssl/openssl.cnf. This was non-hermetic and failed in test environments in
which access to /etc was sandboxed. This patch provides an explicit minimal openssl.cnf, derived
from the /etc/ssl/openssl.cnf on my workstation.