ci: update do_ci.sh targets and docs to the bazel test matrix.
We've lost Mac OS X build information for do_ci.sh, since it's stale and I don't have a way to
verify it. Feel free to add in updated info later, but I think it might be better to use Bazel
directly here instead of via the container build.

Also, the Alpine build is going to atrophy. I'll take an action item to chat with folks about
updating this.