bazel: fix git SHA embedding
Previously, the SHA generation genrule used a glob over .git/**
to figure out if anything changed. This does not work when envoy is
embedded in another git repo as a submodule, or is potentially built
from a tarball. This commit reworks the generation to expect an env
var called GIT_BUILD_SHA which will be used as the final SHA in the
build. CI has been updated to use the GIT SHA being built from.

If the env var is not defined, the SHA is set to a default SHA.

This commit also removes a dependency on version_lib from test main,
which was causing all tests to be relinked for no reason.