Backtrace on fatal signals. (#855)
This commit adds signal handlers into main.cc, or any smaller scope as
desired.  For any fatal signals a backtrace is logged and then the
signal is re-raised to continue with the default behavior.

The backtrace logging and script are modified slightly to support all
objects found and not make assumptions about the first stack frame being
in the binary we care about. This is needed for reasonable output in the
case of abort() or any fatal signal like SEGV that crops up inside a
library function we call.