Begin to support Ipv6 only or Ipv4 only testing (#861)
This PR allows testing in environments that either don't support IPv4 or IPv6 networking. The way this is achieved is by setting an environment variable (ENVOY_IP_TEST_VERSIONS) in test/main.cc or externally to "v4only" or "v6only". We then construct parameterized gtests based on the value of ENVOY_IP_TEST_VERSIONS.

This PR only modifies existing tests that were testing with both IPv4 and IPv6 addresses. If the approach taken in this PR is agreed upon, testing with IPv4 only and IPv6 only will be extended in subsequent PRs.

This PR is related to #351