network: ability to pass in DNS resolvers to DnsResolver
This change allows DNS resolvers to be set for strict_dns and
logical_dns cluster types.

To enable this, DNS resolvers are passed in as an array of ip:port
strings, inside of a "dns_resolvers" setting inside of the cluster
config.

To make this possible, the Strict DNS and Logical DNS implementations
can now be passed in a custom Network::DnsResolver.

Implements https://github.com/lyft/envoy/issues/940.