Add backtrace-on-fatal behavior for tests. (#896)
Instantiate SignalAction at the top level of test/main.cc so fatal
signals occurring in tests will result in logged backtrace.  Use
--run_under=`pwd`/tools/stack_decode.py to decode symbols for quick
debugging.  As with exe/main.cc the feature can be disabled with
--define=signal_trace=disabled.

In the process of testing this the need for restoration of the previous
signal handler was seen, particularly in the case where SignalAction was
explicitly instantiated at a tighter scope inside a test.  SignalAction
is modified to restore the prior signal handler and signal handling
stack upon destruction. A new test verifies this functionality.