Fail nice and hard if old GCC encountered. (#901)
If GCC prior to 4.9 was used to compile the result was a mysterious
runtime failure of 14 different tests that rely on the use of
std::regex but are otherwise unremarkable.  This is because GCC 4.8 and
below have a silently incomplete and buggy std::regex implementation. It
compiles, but it doesn't actually do the right thing at runtime (!!).

To prevent a repeat of the confusion this caused #error out if GCC that
doesn't support regex properly is used to compile
test/test_common/environment.cc.