Fix segfault caused by modifying list within iteration.
A segfault arose when iterating over the list of
secondary_init_clusters_ and calling initialize().  The call to
initialize may result in the initialized item being removed from the
secondary_init_clusters_ list while walking through the list.  The
initialized item can only remove itself, not any other item in the
list.  However, this resulted in chasing an invalidated iterator when
trying to go to the next element in the list.

To fix this the loop is modified so that the iterator is advanced to the
next item (or end() if no more items) before calling
Cluster::initialize().

A test is included which exercises this scenario

Fixes #903