statsd: add connection overlow handling (#930)
We have seen a production issue where the local statsd relay
stops responding without breaking the connection. Eventually this
leads to an OOM condition. This commit adds a safeguard around the
maximum amount of memory that can be buffered before we give up.
In that case we increment a stat that can at least be viewed locally.

Fixes https://github.com/lyft/envoy/issues/897