http filter: remove reset callback and introduce onDestroy()
Previously, if an intermediate filter ends a stream without need
of a reset, further filters would have no way of knowing that and
could not clean up properly. This commit is a breaking change to
the filter API (though trivial in most cases). We remove filter
reset callback registration, and replace it with an onDestroy()
method that is called in all cases. Filters must clean up all
async processing in the onDestroy() callback. This approach is
simpler to understand but also has the benefit of increasing
performance because we no longer have to manage a doubly linked
list and allocate space for the callbacks.

Fixes https://github.com/lyft/envoy/issues/947