http: allow body data to be added be filters during trailers callbacks (#959)
This is needed for GRPC <-> JSON transcoding. See comments in filter.h
for more details. This commit also includes a partial cleanup and hardening
of the connection manager tests.

Note that this is a breaking change for filters that were previously using
the writable buffered data callbacks. They have been replace with a more
robust method.

Fixes https://github.com/lyft/envoy/issues/925