Remove MAX_STREAMS check on client side connection pool. (#963)
The reason for this is that we already have protection via circuit breakers.
There is no point in having additional protection (other than potentially
for push, which we don't support currently, and can deal with later). If we
overflow the other side, the code will correctly handle the REFUSED_STREAM
reset that happens.