build: Disable libunwind in gperftools config. (#1071)
gperftools uses libunwind if the build system has it, but envoy
build system is unaware of this dependency, which causes a build
failure due to libunwind not being included in the link.  On the
other hand, linking libunwind unconditionally causes a link
failure in build environments without libunwind.

gperftools can decode stack traces without libunwind if stack frame
pointers are used.  We already do that for Envoy, so not using
libunwind is only a problem for collectiong stack traces system and
compiler libraries compiled without stack frame pointers.

Disable libunwind in gperftools configuration for now, and revisit
later if needed.

Signed-off-by: Jarno Rajahalme <jarno@covalent.io>