Converted all logs to macros and modified existing log macros. (#1068)
Three types of logging: facility, general, and to a specific logger that the user passes in. I attempted to clean up the way that log macros were being done to make them a bit simpler. They should all filter down to the log_[level] macros for ease of compiling out certain log levels and global modification of log formatting or functionality.