Add generic dispatched thread implementation.
There is a need for generic thread with the "dispatched nature" (ie, it
runs an event loop) for AccessLog writing thread and for other ancillary
threads which will issue gRPC calls.  Like Worker this thread has a
ConnectionHandler and its Dispatcher. Unlike Worker this thread does not
set up listeners (although a custom use that desires this such as for
accepting gRPC calls can be implemented by extending this class).