connection handler: add listener removal support
This commit adds the ability for a listener to be removed from a
connection handler. This functionality is not currently used but
will be used as part of LDS. As part of this change, connections
are now owned by the listener, so that when a listener is destroyed,
all connections are closed. Logic has been cleaned up so that
destroying the connection handler destroys all listeners which
close all connections.